Split templates into two entry classes
